+++ 03-UserManagementWebApp/src/main/java/com/mini/project/service/IUserManagementService.java
@@ -13,9 +13,9 @@ public interface IUserManagementService {
 	
 	public Map<Long, String> getCountries();
 
-	public Map<Long, String> getStates(Integer countryId);
+	public Map<Long, String> getStates(Long countryId);
 
-	public Map<Long, String> getCities(Integer stateId);
+	public Map<Long, String> getCities(Long stateId);
 
 	public String registerUser(UserForm userForm);
